feat: add function to update last_fetch after every API call
