+++ callyzer/callyzer/utils.py
@@ -1,5 +1,3 @@
-
-
 import frappe
 from datetime import datetime
 from frappe import _
@@ -69,4 +67,20 @@ def get_endpoint(endpoint_name):
             break
         
     return endpoint_url, call_from, call_to
+
+def update_last_fetched_time(endpoint_name):
+    found = False
+    settings = frappe.get_single("Callyzer Endpoint Settings")
+
+    for row in settings.endpoints:
+        if row.endpoint_name == endpoint_name:
+            row.last_fetch = frappe.utils.nowdate()
+            found = True
+            break
+
+    if found:
+        settings.save(ignore_permissions=True)
+        return True
+    else:
+        frappe.throw("{endpoint_name} not found in settings")
